company admin and super admin fixes

+++ app/Http/Controllers/Employer/ResourceController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\Company;
 use App\Models\CompanyEmployee;
 use App\Models\CompanyResource;
+use App\Models\Resource;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
@@ -20,12 +21,6 @@ class ResourceController extends Controller
     {
         $user = Auth::guard('api')->user();
         $company = Company::where('user_id', $user->id)->first();
-        if (!$request->company) {
-            $company_id = $company->id;
-        } else {
-            $company_id = $request->company;
-        }
-
         $validator = Validator::make($request->all(), [
             'title' => 'required|max:255',
             'description' => 'required',
@@ -45,14 +40,28 @@ class ResourceController extends Controller
                 $uploadedFile->move($destinationPath, $filename);
             }
 
-            $resource = new CompanyResource();
-            $resource->company_id = $company_id;
+            $resource = new Resource();
             $resource->title = $request->title;
             $resource->description = $request->description;
             $resource->file = $filename;
             $resource->link = $request->link;
             $resource->visibility = $request->visibility;
             $resource->save();
+            if ($request->company == "") {
+                $company_id = $company->id;
+                $cw = new CompanyResource();
+                $cw->company_id = $company_id;
+                $cw->resource_id = $resource->id;
+                $cw->save();
+            } else {
+                $company = json_decode($request->company);
+                foreach ($company as $value) {
+                    $cw = new CompanyResource();
+                    $cw->company_id = $value->id;
+                    $cw->resource_id = $resource->id;
+                    $cw->save();
+                }
+            }
             return response(["status" => "success", 'res' => $resource], 200);
         }
     }
@@ -74,18 +83,17 @@ class ResourceController extends Controller
             $error = $validator->getMessageBag()->first();
             return response()->json(["status" => "error", "message" => $error], 400);
         } else {
-            $resource = CompanyResource::find($request->id);
+            $resource = Resource::find($request->id);
             $filename = $resource->file;
             if ($request->hasFile('file')) {
                 $uploadedFile = $request->file('file');
                 $filename = time() . '_' . $uploadedFile->getClientOriginalName();
                 $destinationPath = public_path() . '/company-resources';
-                unlink($destinationPath . '/' . $resource->file);
+                if($resource->file) {
+                    unlink($destinationPath . '/' . $resource->file);
+                }
                 $uploadedFile->move($destinationPath, $filename);
 
-            }
-            if ($request->company) {
-                $resource->company_id = $request->company;
             }
             $resource->title = $request->title;
             $resource->description = $request->description;
@@ -93,6 +101,16 @@ class ResourceController extends Controller
             $resource->link = $request->link;
             $resource->visibility = $request->visibility;
             $resource->save();
+            if ($request->company) {
+                CompanyResource::where('resource_id', $request->id)->delete();
+                $company = json_decode($request->company);
+                foreach ($company as $value) {
+                    $cw = new CompanyResource();
+                    $cw->company_id = $value->id;
+                    $cw->resource_id = $resource->id;
+                    $cw->save();
+                }
+            }
             return response(["status" => "success", 'res' => $resource], 200);
         }
     }
@@ -109,12 +127,15 @@ class ResourceController extends Controller
         $user = Auth::guard('api')->user();
         $company = CompanyEmployee::where('user_id', $user->id)->first();
         if ($company) {
-            $resources = CompanyResource::where("company_id", $company->company_id);
+            $resources = Resource::select('resources.*')
+                ->join('company_resources','company_resources.resource_id','resources.id')
+                ->where("company_id", $company->company_id);
         } else {
-            $resources = CompanyResource::select('company_resources.*', 'companies.company_name')
-                ->join('companies', 'companies.id', 'company_resources.company_id');
+            $resources = Resource::with(['company' => function ($q) {
+                $q->join('companies', 'companies.id', 'company_resources.company_id')->pluck('companies.company_name');
+            }])->select('resources.*');
         }
-        $path = url('/public/welcome-notes/');
+        $path = url('/public/company-resources/');
         if ($keyword) {
             $resources = $resources->where('title', 'like', "%$keyword%");
         }
@@ -131,8 +152,12 @@ class ResourceController extends Controller
      */
     public function get_resource($id)
     {
-        $resources = CompanyResource::find($id);
-        $resources->file = url('/public/welcome-notes/') . '/' . $resources->file;
+        $resources = Resource::find($id);
+        $resources->file = url('/public/company-resources/') . '/' . $resources->file;
+        $resources->company = CompanyResource::join('companies', 'companies.id', 'company_resources.company_id')
+            ->select('companies.id', 'companies.company_name as name')
+            ->where('resource_id', $id)
+            ->get();
         return response(["status" => "success", 'res' => $resources], 200);
     }
 
@@ -142,12 +167,13 @@ class ResourceController extends Controller
      */
     public function delete_resource($id)
     {
-        $resource = CompanyResource::find($id);
+        $resource = Resource::find($id);
         if ($resource->file) {
             $destinationPath = public_path() . '/company-resources';
             unlink($destinationPath . '/' . $resource->file);
         }
         $resource->delete();
+        CompanyResource::where('resource_id',$id)->delete();
         return response(["status" => "success", 'res' => $resource], 200);
     }
 
@@ -157,7 +183,7 @@ class ResourceController extends Controller
      */
     public function download_file($id)
     {
-        $filename = CompanyResource::find($id);
+        $filename = Resource::find($id);
         $filename = $filename->file;
         $file = public_path() . '/company-resources/' . $filename;
         return response()->download($file);
@@ -170,10 +196,12 @@ class ResourceController extends Controller
     {
         $user = Auth::guard('api')->user();
         $company = CompanyEmployee::where('user_id', $user->id)->first();
-        $resources = CompanyResource::where("company_id", $company->company_id)
-                                    ->orderby("id", "desc")
-                                    ->limit(5)
-                                    ->get();
+        $resources = Resource::select('resources.*')
+                                ->join('company_resources','company_resources.resource_id','resources.id')
+                                ->where("company_id", $company->company_id)
+                                ->orderby("id", "desc")
+                                ->limit(3)
+                                ->get();
         return response(["status" => "success", 'res' => $resources], 200);
     }
